% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_token.R
\name{ices_token}
\alias{ices_token}
\title{ices_token}
\usage{
ices_token(username = NULL, password = NULL, refresh = FALSE, ...)
}
\arguments{
\item{username}{the ices username that you require a token for,
Default: NULL, which results in \code{getOption("ices.username")}}

\item{password}{the ices username that you require a token for,
Default: NULL, which results in a dialogue box request for the
password}

\item{refresh}{should the token be refreshed to update new claims
for example}

\item{...}{not yet used}
}
\value{
character scalar, the token
}
\description{
Stores a user token in the system keyring
}
\examples{
\dontrun{
ices_token()
}

}
