% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_token.R
\name{clear_token}
\alias{clear_token}
\title{clear_token}
\usage{
clear_token(username = getOption("ices.username"))
}
\arguments{
\item{username}{the ices username that you require a token for,
Default: getOption("ices.username")}
}
\value{
no return value
}
\description{
Deletes the token from the key ring.
}
\examples{
\dontrun{
clear_token()
}

}
\seealso{
\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}
}
