% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_token.R
\name{decode_token}
\alias{decode_token}
\title{Decode token}
\usage{
decode_token(jwt)
}
\arguments{
\item{jwt}{a javascript web token got by running \code{ices_token()}}
}
\value{
list of claims
}
\description{
Decodes the token to allow inspection of claims
}
\examples{
\dontrun{
jwt <- ices_token()
decode_token(jwt)
}

}
\seealso{
\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}
}
