% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_token.R, R/token_expiration.R,
%   R/token_user.R
\name{decode_token}
\alias{decode_token}
\alias{token_expiration}
\alias{token_user}
\title{Decode token}
\usage{
decode_token(jwt = ices_token(), formatted = TRUE)

token_expiration(jwt = ices_token())

token_user(jwt = ices_token())
}
\arguments{
\item{jwt}{a javascript web token got by running \code{ices_token()}}
}
\value{
list of claims

POSIXct date time

string, user name
}
\description{
Decodes the token to allow inspection of claims

Return the expiration date of the token

What user is this token for
}
\examples{
\dontrun{
decode_token()
}

\dontrun{
token_expiration()
}

\dontrun{
token_user()
}

}
\seealso{
\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}

\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}

\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}
}
