% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_get_jwt.R
\name{ices_get_jwt}
\alias{ices_get_jwt}
\title{ices_get_jwt}
\usage{
ices_get_jwt(url, retry = TRUE, quiet = FALSE)
}
\arguments{
\item{url}{PARAM_DESCRIPTION}

\item{retry}{PARAM_DESCRIPTION, Default: TRUE}

\item{quiet}{should the request be made queitly}
}
\value{
httr response object
}
\description{
Perform a get request to a url using a javascript web token for authentication
}
\examples{
ices_get_jwt("https://taf.ices.dk/vms/api/gearwidths")

}
\seealso{
\code{\link[icesConnect]{ices_token}}
\code{\link[httr]{GET}},\code{\link[httr]{add_headers}},\code{\link[httr]{stop_for_status}},\code{\link[httr]{status_code}}
}
