% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_token.R
\name{ices_token}
\alias{ices_token}
\title{ices_token}
\usage{
ices_token(...)
}
\arguments{
\item{...}{not yet used}
}
\value{
character scalar, the token
}
\description{
Stores a user token in the system keyring
}
\examples{
\dontrun{
ices_token()
}

}
