% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_token.R
\name{ices_token}
\alias{ices_token}
\title{ices_token}
\usage{
ices_token(username = NULL, password = NULL, ...)
}
\arguments{
\item{username}{the ices username that you require a token for,
Default: NULL, which results in getOption("ices.username")}

\item{password}{the ices username that you require a token for,
Default: NULL, which results in a dialogue box request for the
password}

\item{...}{not yet used}
}
\value{
character scalar, the token
}
\description{
Stores a user token in the system keyring
}
\examples{
\dontrun{
ices_token()
}

}
\seealso{
\code{\link[whoami]{username}}
\code{\link[keyring]{key_get}}
}
