% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_username.R
\name{set_username}
\alias{set_username}
\title{Set your ICES username}
\usage{
set_username(username)
}
\arguments{
\item{username}{the ices username you use to login to SharePoint
for example.}
}
\value{
invisible value: the previous username that was set
}
\description{
Set the ICES user or logon name that will be used for requests
}
\examples{
\dontrun{
ices_token()
}

}
\seealso{
\code{\link[base]{options}}
\code{\link[base]{getOption}}
}
