% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_username.R
\name{set_username}
\alias{set_username}
\title{Set your ICES username}
\usage{
set_username(...)
}
\arguments{
\item{...}{only here for back compatability, does nothing}
}
\value{
invisible NULL
}
\description{
Back compatability function that does nothing.
}
