% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_username.R
\name{set_username}
\alias{set_username}
\title{Set your ICES username}
\usage{
set_username(username)
}
\arguments{
\item{username}{the ices username you use to login to SharePoint
for example.}
}
\value{
invisible value: the previous username that was set
}
\description{
Set the ICES user or logon name that will be used for requests
}
\details{
To set this permanently, add the line:
\code{options(ices.username = "your ices user name")}
to your RProfile file, see, usually '~/.Rprofile', see
\code{\link[base]{Startup}} for more information.
}
\examples{
\dontrun{
set_username("myusername")
}

}
\seealso{
\code{\link[base]{options}}
\code{\link[base]{getOption}}
}
