% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_username.R, R/get_username.R,
%   R/set_username.R
\name{clear_username}
\alias{clear_username}
\alias{get_username}
\alias{set_username}
\title{get_username}
\usage{
clear_username()

get_username()

set_username(username)
}
\arguments{
\item{username}{the ices username you use to login to the ICES SharePoint
for example.}
}
\value{
character scalar, the username

invisible value: the previous username that was set
}
\description{
Get your ICES username from the options.

Set the ICES user or logon name that will be used for requests
}
\examples{
\dontrun{
get_username()
}

\dontrun{
set_username("myusername")
}

}
