% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_username.R, R/reset_username.R,
%   R/set_username.R
\name{get_username}
\alias{get_username}
\alias{reset_username}
\alias{set_username}
\title{get_username}
\usage{
get_username()

reset_username()

set_username(username)
}
\arguments{
\item{username}{the ices username you use to login to SharePoint
for example.}
}
\value{
character scalar, the username

invisible value: the previous username that was set
}
\description{
Get your ICES username from the options.

Set the ICES user or logon name that will be used for requests
}
\details{
To set this permanently, add the line:
\code{options(ices.username = "your ices user name")}
to your '.Rprofile' file, see, usually '~/.Rprofile', see
\code{\link[base]{Startup}} for more information.
}
\examples{
\dontrun{
get_username()
}

\donttest{
set_username("myusername")
}

}
\seealso{
\code{\link[base]{options}}
\code{\link[base]{getOption}}
}
